--- src/components/about/About.tsx
@@ -1,7 +1,0 @@
-export function About() {
-  return (
-    <>
-      <h1>About</h1>
-    </>
-  )
-}
