+++ src/components/generator/form/sections/ProfileSection.tsx
@@ -1,6 +1,6 @@
 // import { useFieldArray } from 'react-hook-form'
 import { FormSection } from './FormSection'
-import { LabeledInput } from '../inputs/LabeledInput'
+import { LabeledInput } from '../../../core/LabeledInput'
 
 export function ProfileSection() {
   // const { fields, append } = useFieldArray({ name: 'basics.links' })
