+++ src/components/generator/layout/Header.tsx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import styled from 'styled-components'
 
-import { Logo } from '../../common/Logo'
+import { Logo } from '../../core/Logo'
 import { colors } from '../../../theme'
 
 const StyledHeader = styled.header`
