+++ src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app'
 import Head from 'next/head'
-import { GlobalStyle } from '../components/common/GlobalStyle'
+import { GlobalStyle } from '../components/core/GlobalStyle'
 
 function App({ Component, pageProps }: AppProps) {
   return (
